+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.3.0] - 2025-01-25
+* change FxHashMap to AHashMap
+* change id_ptr: &'static IdCoreNxM<N, M> to id_ptr: *const IdCoreNxM<N, M>,
+
 ## [0.2.2] - 2024-09-26
 * added structs: 
   * StaticId16x0